extract link from text

+++ bot.py
@@ -9,25 +9,33 @@ from datetime import datetime
 from pinterest_caller import download_link
 from pathlib import Path
 import shutil
+import re
 
 def start(update, context):
     bot = context.bot
     log_text("start " + str(update.message.chat_id) + " " + str(update.message.from_user.username) + " " + str(update.message.chat.title), bot)
+
+def find_link(text):
+    regex=ur"\b((?:https?://)?(?:(?:www\.)?(?:[\da-z\.-]+)\.(?:[a-z]{2,6})|(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)|(?:(?:[0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|(?:[0-9a-fA-F]{1,4}:){1,7}:|(?:[0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|(?:[0-9a-fA-F]{1,4}:){1,5}(?::[0-9a-fA-F]{1,4}){1,2}|(?:[0-9a-fA-F]{1,4}:){1,4}(?::[0-9a-fA-F]{1,4}){1,3}|(?:[0-9a-fA-F]{1,4}:){1,3}(?::[0-9a-fA-F]{1,4}){1,4}|(?:[0-9a-fA-F]{1,4}:){1,2}(?::[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:(?:(?::[0-9a-fA-F]{1,4}){1,6})|:(?:(?::[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(?::[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(?:ffff(?::0{1,4}){0,1}:){0,1}(?:(?:25[0-5]|(?:2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(?:25[0-5]|(?:2[0-4]|1{0,1}[0-9]){0,1}[0-9])|(?:[0-9a-fA-F]{1,4}:){1,4}:(?:(?:25[0-5]|(?:2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(?:25[0-5]|(?:2[0-4]|1{0,1}[0-9]){0,1}[0-9])))(?::[0-9]{1,4}|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-5])?(?:/[\w\.-]*)*/?)\b"
+    matches = re.findall(regex, text)
+    return matches[0] if len(matches) else None
 
 def free_text(update, context):
     bot = context.bot
     laptop_image_path = random.choice(glob.glob("./laptop_images/*"))
     with open(laptop_image_path, 'rb') as f:
         bot.send_sticker(update.message.chat_id, f)
-    file_paths, download_dir = download_link(update.message.text)
-    # log_text("file_paths: " + str(file_paths), bot)
-    for file_path in file_paths:
-        # log_text("sending: " + str(file_path), bot)
-        with open(file_path, 'rb') as f:
-            bot.send_document(update.message.chat_id, document=f, filename=Path(file_path).name)
-    # log_text("removing: " + download_dir, bot)
-    shutil.rmtree(download_dir)
-    # log_text("removed: " + download_dir, bot)
+    link = find_link(update.message.text)
+    if link:
+        file_paths, download_dir = download_link(link)
+        # log_text("file_paths: " + str(file_paths), bot)
+        for file_path in file_paths:
+            # log_text("sending: " + str(file_path), bot)
+            with open(file_path, 'rb') as f:
+                bot.send_document(update.message.chat_id, document=f, filename=Path(file_path).name)
+        # log_text("removing: " + download_dir, bot)
+        shutil.rmtree(download_dir)
+        # log_text("removed: " + download_dir, bot)
 
 def echo(update, context):
     bot = context.bot
